requirements.txt
langchain-google-genai
langchain-core
langchain-openai
langchain-community
langchainhub

python-dotenv
streamlit

beautifulsoup4
pypdf
sentence-transformers
wikipedia
arxiv
